Extract project generator rules for MCP reuse
